requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
face-recognition==1.3.0
fer==22.4.0
Pillow==9.5.0
python-dotenv==1.0.0
pytest==7.3.1
Flask-Swagger-UI==4.11.1
Werkzeug==2.2.3 